% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid-package.R
\docType{package}
\name{covid-package}
\alias{covid}
\alias{covid-package}
\title{covid: COVID Tracking Project API wrapper}
\description{
This package provides data on US cases of
    COVID-19.
}
\author{
\strong{Maintainer}: Amanda Dobbyn \email{amanda.e.dobbyn@gmail.com}

}
\keyword{internal}
