% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_counties_info}
\alias{get_counties_info}
\title{Get COVID-related information for certain counties}
\usage{
get_counties_info()
}
\value{
A tibble with one row per county and their COVID website information.
}
\description{
Currently limited to the worst-affected counties in mostly Washington state, California, and New York.
}
\examples{
\dontrun{
get_counties_info()
}
}
