% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_info_covid19us}
\alias{get_info_covid19us}
\title{Get info about this dataset}
\usage{
get_info_covid19us()
}
\value{
A tibble with information about where the data is pulled from, details about the dataset, what the data types are, etc.
}
\description{
Get info about this dataset
}
\examples{
\donttest{
get_info_covid19us()
}
}
