% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_states_daily}
\alias{get_states_daily}
\title{Get daily counts for every state}
\usage{
get_states_daily()
}
\value{

}
\description{
Updated every day at 4pm.
}
\examples{
\dontrun{
get_states_daily()
}
}
