% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_states_info}
\alias{get_states_info}
\title{Get COVID-related information for each state}
\usage{
get_states_info()
}
\value{
A tibble with one row per state incluing information on the state's \code{data_site} where the data was pulled from and the \code{covid_19_site} where data is published.
}
\description{
Get COVID-related information for each state
}
\examples{
\dontrun{
get_states_info()
}
}
