% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_tracker_urls}
\alias{get_tracker_urls}
\title{Get URLs and their details for each state}
\usage{
get_tracker_urls()
}
\value{
A tibble with one row for every state, the URL used by scrapers to get data, and a \code{filter} column that provices the xpath or CSS selector used by the \href{https://github.com/COVID19Tracking/covid-tracking}{COVID-19 Tracking Project's scraper} to get this data.
}
\description{
Get URLs and their details for each state
}
\examples{
\donttest{
get_tracker_urls()
}
}
