% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_us_current}
\alias{get_us_current}
\title{Get current US counts}
\usage{
get_us_current()
}
\value{
A tibble with one row for the current count of the country's COVID statuses.
}
\description{
Get current US counts
}
\examples{
\dontrun{
get_us_current()
}
}
