% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_us_daily}
\alias{get_us_daily}
\title{Get daily US counts}
\usage{
get_us_daily()
}
\value{
A tibble with one row per date in which data is available and counts for each of those states.
}
\description{
Updated every day at 4pm.
}
\examples{
\dontrun{
get_us_daily()
}
}
