% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{refresh_covid19us}
\alias{refresh_covid19us}
\title{Get state data in long format}
\usage{
refresh_covid19us(type = "daily")
}
\arguments{
\item{type}{One of \code{"daily"} or \code{"current"}}
}
\value{
A tibble of data retrieved with \code{\link{get_states_daily}} or \code{\link{get_states_current}} in long format with a \code{data_type} and a \code{value} column.
}
\description{
Get state data in long format
}
\examples{
\donttest{
refresh_covid19us()
}
}
