% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19us-package.R
\docType{package}
\name{covid19us-package}
\alias{covid19us}
\alias{covid19us-package}
\title{covid19us: Cases of COVID-19 in the United States}
\description{
This package provides data on US cases of
    COVID-19 from the COVID Tracking Project API <https://covidtracking.com/api/>.
}
\author{
\strong{Maintainer}: Amanda Dobbyn \email{amanda.e.dobbyn@gmail.com}

}
\keyword{internal}
