% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_states_current}
\alias{get_states_current}
\title{Get current counts for every state}
\usage{
get_states_current()
}
\value{

}
\description{
Get current counts for every state
}
\examples{
\dontrun{
get_states_current()
}
}
