% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_tracker_urls}
\alias{get_tracker_urls}
\title{Get URLs and their details for each state}
\usage{
get_tracker_urls()
}
\value{

}
\description{
Includes the xpath css selector used to get data.
}
\examples{
\dontrun{
get_tracker_urls()
}
}
